requirements.txt
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
Flask==2.0.1
future==0.18.2
google-auth==2.1.0
google-auth-oauthlib==0.4.6
gspread==4.0.1
gunicorn==20.1.0
httplib2==0.19.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
line-bot-sdk==1.20.0
MarkupSafe==2.0.1
oauth2client==4.1.3
oauthlib==3.1.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dotenv==0.19.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
urllib3==1.26.6
Werkzeug==2.0.1
